requirement.txt
django
djangorestframework
markdown
django-filter
django-cleanup
mysqlclient
pillow
google-generativeai
django-cors-headers
scikit-learn
djangorestframework-simplejwt